fix macOS build after bugfix bundle merge
- guard Windows-only hard-coded zlib path in src/CMakeLists.txt
- fswindow.cpp: wtypes.h/GetDesktopWindow only on Windows
- fscontrol.cpp: Sleep -> FsSleep
- fsnetwork.cpp: pass YsString as C string to printf
- fsgroundpropertygl2.0.cpp: explicit float casts

+++ src/graphics/gl2.0/fsgroundpropertygl2.0.cpp
@@ -16,7 +16,7 @@ void FsAircraftCarrierProperty::EndDrawArrestingWire(void)
 	YsGLSL3DRenderer *renderer=YsGLSLSharedFlat3DRenderer();
 	YsGLSLUse3DRenderer(renderer);
 
-	GLfloat color[4]={r,g,b,1.0f};
+	GLfloat color[4]={(GLfloat)r,(GLfloat)g,(GLfloat)b,1.0f};
 	YsGLSLSet3DRendererUniformColorfv(renderer,color);
 
 	YsGLSLDrawPrimitiveVtxfv(renderer,GL_LINES,vtxBuf.nVtx,vtxBuf.vtxArray);
